add-course work need edit-course

+++ client/src/app/services/courses.service.ts
@@ -1,36 +1,48 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Course } from '../models/course';
+import { Course } from '../models/course.model';
 
 @Injectable({
   providedIn: 'root',
 })
 export class CoursesService {
-  constructor(private http: HttpClient) {}
+  constructor(private http: HttpClient) { }
 
   getCourses(): Observable<Course[]> {
     const token = localStorage.getItem('token');
 
-    const headers = new HttpHeaders({
-      Authorization: `Bearer ${token}`,
-    });
+    if (token) {
+      const headers = new HttpHeaders({
+        Authorization: `Bearer ${token}`,
+      });
 
-    return this.http.get<Course[]>('http://localhost:3000/api/courses', {headers});
+      return this.http.get<Course[]>('http://localhost:3000/api/courses', { headers });
+    }
+
+    throw new Error("No token found");
   }
 
   getCourseById(id: number): Observable<Course> {
     const token = localStorage.getItem('token');
 
+    if (!token) {
+      throw new Error("No token found");
+    }
+
     const headers = new HttpHeaders({
       Authorization: `Bearer ${token}`,
     });
 
-    return this.http.get<Course>(`http://localhost:3000/api/courses/:${id}`, {headers});
+    return this.http.get<Course>(`http://localhost:3000/api/courses/${id}`, { headers });
   }
 
   addCourse(course: Course): Observable<any> {
     const token = localStorage.getItem('token');
+
+    if (!token) {
+      throw new Error("No token found");
+    }
 
     const headers = new HttpHeaders({
       Authorization: `Bearer ${token}`,
@@ -38,32 +50,41 @@ export class CoursesService {
 
     return this.http.post<Course>(
       `http://localhost:3000/api/courses`,
-      course, {headers}
+      course, { headers }
     );
   }
 
-  updateCourse(id:number, course: Course): Observable<any> {
+  updateCourse(id: number, course: Course): Observable<any> {
     const token = localStorage.getItem('token');
+
+    if (!token) {
+      throw new Error("No token found");
+    }
 
     const headers = new HttpHeaders({
       Authorization: `Bearer ${token}`,
     });
 
     return this.http.put<Course>(
-      `http://localhost:3000/api/courses/:${id}`,
-      course, {headers}
+      `http://localhost:3000/api/courses/${id}`,
+      course, { headers }
     );
   }
 
   deleteCourse(id: number): Observable<any> {
     const token = localStorage.getItem('token');
 
+    if (!token) {
+      throw new Error("No token found");
+    }
+
     const headers = new HttpHeaders({
       Authorization: `Bearer ${token}`,
     });
 
     return this.http.delete<Course>(
-      `http://localhost:3000/api/courses/:${id}`, {headers}
+      `http://localhost:3000/api/courses/${id}`, { headers }
     );
   }
 }
+
